Create Number of Factors
Alice has learnt factorization recently. Bob doesn't think she has learnt it properly and hence he has decided to quiz her. Bob gives Alice a very large number and asks her to find out the number of factors of that number. To make it a little easier for her, he represents the number as a product of N numbers. Alice is frightened of big numbers and hence is asking you for help. Your task is simple. Given N numbers, you need to tell the number of distinct factors of the product of these N numbers.

Example:

Input:
3       //number of test cases
3       //number of numbers in first test case
3  5  7      //numbers that when all multiplied gives the number whose no.of factors are to be found
3
2  4  6
2
5  5

Output:
8
10
3